+++ common/src/main/java/net/mehvahdjukaar/polytone/mixins/EnvironmentSystemMixin.java
@@ -1,6 +1,8 @@
 package net.mehvahdjukaar.polytone.mixins;
 
 import net.mehvahdjukaar.polytone.Polytone;
+import net.minecraft.world.attribute.EnvironmentAttribute;
+import net.minecraft.world.attribute.EnvironmentAttributeMap;
 import net.minecraft.world.attribute.EnvironmentAttributeSystem;
 import net.minecraft.world.level.Level;
 import org.spongepowered.asm.mixin.Mixin;
@@ -16,4 +18,6 @@ public class EnvironmentSystemMixin {
         Polytone.BIOME_MODIFIERS.addPostLayers(builder, level);
         Polytone.DIMENSION_MODIFIERS.addPostLayers(builder, level);
     }
+
+
 }